Move version comparison to separate package
